refactors census atlas component and function to implement census atlas route
